corrected error

+++ tofu/geom/_vignetting_tools.pxd
@@ -67,4 +67,4 @@ cdef inline int are_in_vignette(int sz_r, int sz_z,
                                 int npts_vpoly,
                                 double* disc_r,
                                 double* disc_z,
-                                double[:, ::1] is_in_vignette) nogil
+                                long[:, ::1] is_in_vignette) nogil
